feat: introduce DirectoryHandler in template; refactor outputs and metrics setup; introduce FieldDetectors and FieldDetections; refactor read_omr_response; abstract bubbles thresholding into its own class; make way for ocr, qr and more detectors; wip
