+++ src/algorithm/detection/ocr/detector.py
@@ -1,0 +1,6 @@
+from src.algorithm.detection.field import FieldTypeDetector
+
+
+class OCRFieldDetector(FieldTypeDetector):
+    def read_field(self, field, gray_image, colored_image, file_aggregate_params):
+        return
