+++ src/algorithm/template/alignment/k_nearest_interpolation.py
