+++ src/algorithm/template/alignment/phase_correlation.py
