+++ src/algorithm/template/alignment/piecewise_affine_delaunay.py
