+++ src/algorithm/template/alignment/sift_matcher.py
