+++ src/algorithm/template/alignment/utils.py
