+++ src/entry.py
@@ -1,12 +1,3 @@
-"""
-
- OMRChecker
-
- Author: Udayraj Deshmukh
- Github: https://github.com/Udayraj123
-
-"""
-
 import json
 import os
 from csv import QUOTE_NONNUMERIC
@@ -17,12 +8,15 @@ import pandas as pd
 from rich.table import Table
 
 from src.algorithm.alignment.template_alignment import apply_template_alignment
-from src.algorithm.evaluation import EvaluationConfig, evaluate_concatenated_response
-from src.algorithm.template import Template
+from src.algorithm.evaluation.config import (
+    EvaluationConfig,
+    evaluate_concatenated_response,
+)
+from src.algorithm.template.template_layout import Template
 from src.schemas.constants import DEFAULT_ANSWERS_SUMMARY_FORMAT_STRING
 from src.schemas.defaults import CONFIG_DEFAULTS
 from src.utils import constants
-from src.utils.file import Paths, setup_dirs_for_paths, setup_outputs_for_template
+from src.utils.file import PathUtils, setup_dirs_for_paths, setup_outputs_for_template
 from src.utils.image import ImageUtils
 from src.utils.interaction import InteractionUtils, Stats
 from src.utils.logger import console, logger
@@ -38,47 +32,6 @@ def entry_point(input_dir, args):
         raise Exception(f"Given input directory does not exist: '{input_dir}'")
     curr_dir = input_dir
     return process_dir(input_dir, curr_dir, args)
-
-
-def export_omr_metrics(
-    outputs_namespace,
-    file_name,
-    image,
-    final_marked,
-    colored_final_marked,
-    template,
-    field_number_to_field_bubble_means,
-    global_threshold_for_template,
-    global_field_confidence_metrics,
-    evaluation_meta,
-):
-    global_bubble_means_and_refs = []
-    for field_bubble_means in field_number_to_field_bubble_means:
-        global_bubble_means_and_refs.extend(field_bubble_means)
-    # sorted_global_bubble_means_and_refs = sorted(global_bubble_means_and_refs)
-
-    image_metrics_path = outputs_namespace.paths.image_metrics_dir.joinpath(
-        f"{os.path.splitext(file_name)[0]}.js"
-    )
-    with open(
-        image_metrics_path,
-        "w",
-    ) as f:
-        json_string = json.dumps(
-            {
-                "global_threshold_for_template": global_threshold_for_template,
-                "template": template,
-                "evaluation_meta": (
-                    evaluation_meta if evaluation_meta is not None else {}
-                ),
-                "global_bubble_means_and_refs": global_bubble_means_and_refs,
-                "global_field_confidence_metrics": global_field_confidence_metrics,
-            },
-            default=lambda x: x.to_json(),
-            indent=4,
-        )
-        f.write(f"export default {json_string}")
-        logger.info(f"Exported image metrics to: {image_metrics_path}")
 
 
 def print_config_summary(
@@ -101,7 +54,7 @@ def print_config_summary(
     table.add_row("Set Layout Mode ", "ON" if args["setLayout"] else "OFF")
     table.add_row(
         "Markers Detection",
-        "ON" if "CropOnMarkers" in template.pre_processors else "OFF",
+        "ON" if "CropOnMarkers" in template.get_pre_processors() else "OFF",
     )
     table.add_row("Detected Template Path", f"{template}")
     if local_config_path:
@@ -111,9 +64,9 @@ def print_config_summary(
 
     table.add_row(
         "Detected pre-processors",
-        f"{[pp.__class__.__name__ for pp in template.pre_processors]}",
+        f"{[pp.__class__.__name__ for pp in template.get_pre_processors()]}",
     )
-    table.add_row("Processing Image Shape", f"{template.processing_image_shape}")
+    table.add_row("Processing Image Shape", f"{template.get_processing_image_shape()}")
 
     console.print(table, justify="center")
 
@@ -142,26 +95,19 @@ def process_dir(
     # Look for subdirectories for processing
     subdirs = [d for d in curr_dir.iterdir() if d.is_dir()]
 
-    output_dir = Path(args["output_dir"], curr_dir.relative_to(root_dir))
     output_mode = args["output_mode"]
-    paths = Paths(output_dir)
 
     # look for images in current dir to process
     exts = ("*.[pP][nN][gG]", "*.[jJ][pP][gG]", "*.[jJ][pP][eE][gG]")
     omr_files = sorted([f for ext in exts for f in curr_dir.glob(ext)])
 
-    # omr_files = Paths.filter_omr_files(omr_files)
+    # omr_files = PathUtils.filter_omr_files(omr_files)
     omr_files = [Path(PurePosixPath(omr_file).as_posix()) for omr_file in omr_files]
 
     # Exclude images (take union over all pre_processors)
     excluded_files = []
     if template:
-        if template.alignment["reference_image_path"] is not None:
-            # Note: reference_image_path is already Path()
-            excluded_files.extend(template.alignment["reference_image_path"])
-
-        for pp in template.pre_processors:
-            excluded_files.extend(Path(p) for p in pp.exclude_files())
+        excluded_files.extend(template.get_exclude_files())
 
     local_evaluation_path = curr_dir.joinpath(constants.EVALUATION_FILENAME)
     # Note: if setLayout is passed, there's no need to load evaluation file
@@ -195,8 +141,10 @@ def process_dir(
                 f"No template file found in the directory tree of {curr_dir}"
             )
 
-        setup_dirs_for_paths(paths)
-        outputs_namespace = setup_outputs_for_template(paths, template, output_mode)
+        output_dir = Path(args["output_dir"], curr_dir.relative_to(root_dir))
+
+        # Reset all mutations to the template, and setup output directories
+        template.reset_and_setup_for_directory(output_dir, output_mode)
 
         print_config_summary(
             curr_dir,
@@ -211,12 +159,11 @@ def process_dir(
             show_template_layouts(omr_files, template, tuning_config)
         else:
             # TODO: pick these args from self/class instead of prop forwarding
-            process_files(
+            process_directory_files(
                 omr_files,
                 template,
                 tuning_config,
                 evaluation_config,
-                outputs_namespace,
                 output_mode,
             )
 
@@ -248,9 +195,7 @@ def show_template_layouts(omr_files, template, tuning_config):
             gray_image,
             colored_image,
             template,
-        ) = template.image_instance_ops.apply_preprocessors(
-            file_path, gray_image, colored_image, template
-        )
+        ) = template.apply_preprocessors(file_path, gray_image, colored_image, template)
         gray_layout, colored_layout = TemplateDrawing.draw_template_layout(
             gray_image,
             colored_image,
@@ -271,12 +216,11 @@ def show_template_layouts(omr_files, template, tuning_config):
         )
 
 
-def process_files(
+def process_directory_files(
     omr_files,
     template,
     tuning_config,
     evaluation_config,
-    outputs_namespace,
     output_mode,
 ):
     start_time = int(time())
@@ -285,7 +229,7 @@ def process_files(
     logger.set_log_levels(tuning_config.outputs.show_logs_by_type)
     for file_path in omr_files:
         files_counter += 1
-        file_name = Paths.remove_non_utf_characters(file_path.name)
+        file_name = PathUtils.remove_non_utf_characters(file_path.name)
         file_id = str(file_name)
 
         gray_image, colored_image = ImageUtils.read_image_util(
@@ -310,9 +254,7 @@ def process_files(
             gray_image,
             colored_image,
             template,
-        ) = template.image_instance_ops.apply_preprocessors(
-            file_path, gray_image, colored_image, template
-        )
+        ) = template.apply_preprocessors(file_path, gray_image, colored_image, template)
 
         # TODO[later]: template as a "Processor"
         # TODO: apply template alignment
@@ -322,21 +264,22 @@ def process_files(
 
         if gray_image is None:
             # Error OMR case
-            output_file_path = outputs_namespace.paths.errors_dir.joinpath(file_name)
-            outputs_namespace.OUTPUT_SET.append(
-                [file_name] + outputs_namespace.empty_resp
-            )
+            output_file_path = template.get_errors_dir().joinpath(file_name)
+
+            template.append_empty_row(file_name)
+
             if check_and_move(
                 constants.ERROR_CODES.NO_MARKER_ERR, file_path, output_file_path
             ):
-                err_line = [
+                error_file_line = [
                     file_name,
                     file_path,
                     output_file_path,
                     "NA",
-                ] + outputs_namespace.empty_resp
-                pd.DataFrame(err_line, dtype=str).T.to_csv(
-                    outputs_namespace.files_obj["Errors"],
+                ] + template.get_empty_response_array()
+
+                pd.DataFrame(error_file_line, dtype=str).T.to_csv(
+                    template.get_errors_file(),
                     mode="a",
                     quoting=QUOTE_NONNUMERIC,
                     header=False,
@@ -344,20 +287,14 @@ def process_files(
                 )
             continue
 
-        (
-            raw_omr_response,
-            multi_marked,
-            _,
-            field_number_to_field_bubble_means,
-            global_threshold_for_template,
-            global_field_confidence_metrics,
-        ) = template.image_instance_ops.read_omr_response(
+        raw_omr_response = template.read_omr_response(
             gray_image, colored_image, template, file_path
         )
 
         # TODO: move inner try catch here
         # concatenate roll nos, set unmarked responses, etc
         concatenated_response = get_concatenated_response(raw_omr_response, template)
+
         evaluation_config_for_response = (
             None
             if evaluation_config is None
@@ -389,8 +326,16 @@ def process_files(
         else:
             logger.info(f"(/{files_counter}) Processed file: '{file_id}'")
 
-        save_marked_dir = outputs_namespace.paths.save_marked_dir
+        # TODO: move this logic inside the class
+        save_marked_dir = template.get_save_marked_dir()
 
+        # TODO: refactor this logic
+        (
+            multi_marked,
+            field_number_to_field_bubble_means,
+            all_fields_threshold_for_file,
+            confidence_metrics_for_file,
+        ) = template.get_metrics_for_file(file_path)
         # Save output image with bubble values and evaluation meta
         if output_mode != "moderation":
             (
@@ -438,18 +383,19 @@ def process_files(
             images_per_row=5 if tuning_config.outputs.show_image_level >= 5 else 4,
         )
 
+        # TODO: move it inside template finalize()
         # Save output metrics
         if tuning_config.outputs.save_image_metrics:
-            export_omr_metrics(
-                outputs_namespace,
+            template.export_omr_metrics_for_file(
                 file_name,
                 gray_image,
                 final_marked,
                 colored_final_marked,
                 template,
+                # TODO: no need to pass these metrics from outside
                 field_number_to_field_bubble_means,
-                global_threshold_for_template,
-                global_field_confidence_metrics,
+                all_fields_threshold_for_file,
+                confidence_metrics_for_file,
                 evaluation_meta,
             )
 
@@ -457,18 +403,17 @@ def process_files(
         output_omr_response = (
             raw_omr_response if output_mode == "moderation" else concatenated_response
         )
-        omr_response_array = []
-        for field in outputs_namespace.omr_response_columns:
-            omr_response_array.append(output_omr_response[field])
+        omr_response_array = template.append_output_omr_response(
+            file_name, output_omr_response
+        )
 
-        outputs_namespace.OUTPUT_SET.append([file_name] + omr_response_array)
-        posix_file_path = Paths.sep_based_posix_path(file_path)
+        posix_file_path = PathUtils.sep_based_posix_path(file_path)
 
         if multi_marked == 0 or not tuning_config.outputs.filter_out_multimarked_files:
             STATS.files_not_moved += 1
 
             # Normalize path and convert to posix style
-            output_file_path = Paths.sep_based_posix_path(
+            output_file_path = PathUtils.sep_based_posix_path(
                 save_marked_dir.joinpath(file_id)
             )
             # Enter into Results sheet-
@@ -481,7 +426,7 @@ def process_files(
 
             # Write/Append to results_line file(opened in append mode)
             pd.DataFrame(results_line, dtype=str).T.to_csv(
-                outputs_namespace.files_obj["Results"],
+                template.get_results_file(),
                 mode="a",
                 quoting=QUOTE_NONNUMERIC,
                 header=False,
@@ -490,8 +435,8 @@ def process_files(
         else:
             # multi_marked file
             logger.info(f"[{files_counter}] Found multi-marked file: '{file_id}'")
-            output_file_path = Paths.sep_based_posix_path(
-                outputs_namespace.paths.multi_marked_dir.joinpath(file_name)
+            output_file_path = PathUtils.sep_based_posix_path(
+                template.get_multi_marked_dir().joinpath(file_name)
             )
             if check_and_move(
                 constants.ERROR_CODES.MULTI_BUBBLE_WARN, file_path, output_file_path
@@ -503,7 +448,7 @@ def process_files(
                     "NA",
                 ] + omr_response_array
                 pd.DataFrame(mm_line, dtype=str).T.to_csv(
-                    outputs_namespace.files_obj["MultiMarked"],
+                    template.get_multi_marked_file(),
                     mode="a",
                     quoting=QUOTE_NONNUMERIC,
                     header=False,
@@ -514,6 +459,10 @@ def process_files(
             #     pass
 
     logger.reset_log_levels()
+
+    # Calculate folder level stats here
+    template.finalize_directory_metrics()
+    # TODO: export directory level stats here
 
     print_stats(start_time, files_counter, tuning_config)
