+++ src/utils/image.py
@@ -2,6 +2,7 @@ import cv2
 import numpy as np
 from matplotlib import pyplot
 from shapely import LineString, Point
+
 from src.processors.constants import EDGE_TYPES_IN_ORDER, EdgeType
 from src.utils.constants import CLR_WHITE
 from src.utils.logger import logger
@@ -49,15 +50,20 @@ class ImageUtils:
     @staticmethod
     def resize_to_shape(img, image_shape):
         h, w = image_shape
-        return ImageUtils.resize_util(img, w, h)
+        return ImageUtils.resize_single(img, w, h)
 
     @staticmethod
-    def resize_to_dimensions(img, image_dimensions):
+    def resize_to_dimensions(image_dimensions, *images):
         w, h = image_dimensions
-        return ImageUtils.resize_util(img, w, h)
+        if len(images) == 1:
+            return ImageUtils.resize_single(images[0], w, h)
+        return map(lambda image: ImageUtils.resize_single(image, w, h), images)
 
-    def resize_util(img, u_width=None, u_height=None):
-        h, w = img.shape[:2]
+    @staticmethod
+    def resize_single(image, u_width=None, u_height=None):
+        if image is None:
+            return None
+        h, w = image.shape[:2]
         if u_height is None:
             u_height = int(h * u_width / w)
         if u_width is None:
@@ -65,8 +71,8 @@ class ImageUtils:
 
         if u_height == h and u_width == w:
             # No need to resize
-            return img
-        return cv2.resize(img, (int(u_width), int(u_height)))
+            return image
+        return cv2.resize(image, (int(u_width), int(u_height)))
 
     @staticmethod
     def get_cropped_warped_rectangle_points(ordered_page_corners):
@@ -132,8 +138,19 @@ class ImageUtils:
         return cnts
 
     @staticmethod
-    def normalize(img, alpha=0, beta=255):
-        return cv2.normalize(img, alpha, beta, norm_type=cv2.NORM_MINMAX)
+    def normalize_single(image, alpha=0, beta=255, norm_type=cv2.NORM_MINMAX):
+        if image is None or image.max() == image.min():
+            return image
+        return cv2.normalize(image, alpha, beta, norm_type)
+
+    @staticmethod
+    def normalize(*images, alpha=0, beta=255, norm_type=cv2.NORM_MINMAX):
+        if len(images) == 1:
+            return ImageUtils.normalize_single(images[0], alpha, beta, norm_type)
+        return map(
+            images,
+            lambda image: ImageUtils.normalize_single(image, alpha, beta, norm_type),
+        )
 
     @staticmethod
     def auto_canny(image, sigma=0.93):
